// packages/ui/exchange-ui-core/src/utils/exchange/types.ts
export interface IAsset {
    id: string;
    address: string;
    tokenId: string;
    deviceId: string;
    generationFrom: string;
    generationTo: string;
}
// ...
export type BundleItem = {
    id: string;
    asset: IAsset;
    startVolume: BigNumber;
    currentVolume: BigNumber;
};

export type Bundle = {
    id?: string;
    userId: string;
    price: number;
    isCancelled: boolean;
    items: BundleItem[];
    volume: BigNumber;
    own: boolean;
    splits?: Split[];
};

export type BundleItemDTO = {
    assetId: string;
    volume: string;
};

export type CreateBundleDTO = {
    price: number;
    items: BundleItemDTO[];
};

export type BuyBundleDTO = {
    bundleId: string;
    volume: string;
};

export type BundleSplits = {
    id: string;
    splits: Split[];
};

export type Split = {
    volume: BigNumber;
    items: SplitItem[];
};

export type SplitItem = {
    id: string;
    volume: BigNumber;
};